Configure the refresh token only
While the client id and secret are pretty stable for something like
this, the refresh_token value might need to change over time if the
scope of that token needs to change. Even for testing, I find it useful
to be able to just issue a new grant token and update the refresh token.